+++ CHANGELOG.md
@@ -21,6 +21,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fixed patient list fragment adapter count
 - Fixed bug on search result on patient list
 - Fixed bug on barcode crash
+- Fixed bug on record vaccine
+- Fixed bug on patient list last seen date
 
 ### Changed
 - Reduce font size and spacing on patient list screen